pyproject.toml
[tool.poetry]
name = "realworlddrf"
version = "0.1.0"
description = ""
authors = ["Sean-Miningah <seanpminingah@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
whitenoise = "^6.5.0"
django-taggit = "^4.0.0"
markdown = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
